ts_load_module: fix comparison with 0 by casting to an integer
This is about building share libraries. countof() static modules will
always be 0 and gcc warns about the comparison always being false due
to limited data range.

This is a bit controversal. I wish I had a better solution now, but we
*do* know that our countof() macro is correct. So I have no problem casting
it to int.

And of course we need i to be signed when comparing it to 0.

+++ src/ts_load_module.c
@@ -95,9 +95,9 @@ static struct tslib_module_info *__ts_load_module_static(struct tsdev *ts,
 							 const char *params)
 {
 	struct tslib_module_info *info = NULL;
-	unsigned int i;
+	int i;
 
-	for (i = 0; i < countof(tslib_modules); i++) {
+	for (i = 0; i < (int)countof(tslib_modules); i++) {
 		if (!strcmp(tslib_modules[i].name, module)) {
 			info = tslib_modules[i].mod_init(ts, params);
 #ifdef DEBUG
